Add functional test simulating a Human v Human game
This commit introduces a logger for debugging.  The logger isn't
configured to print to stdout or a file but pytest is smart enough
to capture log output on failing tests. To show the log output during
passing tests, set the log level as a pytest option:

        $ pytest --log-cli-level=INFO game_test.py

I reformatted game.py and game_test.py with black prior to committing.