Allow players to choose who goes first
Adds a new screen asking who goes first.
Players can also "flip a coin" to decide.

This commit introduces some changes to the logger.
Logging to a file is useful for testing, but it's
not good to have tests and prod log to the same file.

To fix this, the game object can be initialized
with args for logging and for a file to log to.