Refactor curses application views
This commit refactors a single Screen into subclasses for each type of
screen. Instead of `screen.draw_welcome_screen()`, now we create a
`WelcomeScreen` object and call `.draw()` on it.

This refactor allows for better key handling that is specific to
each screen, and not split across Game and Screen methods.

The main reason for this refactor is to allow for better editing
of player settings. The `SettingsScreen` shows settings for
`Human` and `Computer` players. With this refactor, each player's
settings are shown in a subwindow, making the drawing easier.

Adding the subwin forced some changes to the functional tests,
because keypresses need to be mocked from the main application window
and from the subwindows.