Allow setting computer player difficulty
The main Settings screen has two subwindows: one for each player's
settings. This commit cleans up the logic that handles the player
windows specifically as it relates to passing error messages from the
window to the main Settings screen. This way the prompts and error
messages can be shown below the subwindows.